+++ spec/tests/ContentMetadata.spec.ts
@@ -69,7 +69,7 @@ describe('content metadata spec', () => {
       );
     });
 
-    it('override playerType in custom tags', () => {
+    it('do not override playerType in custom tags', () => {
       jest.spyOn(playerMock, 'getPlayerType').mockReturnValue(PlayerType.Native);
       jest.spyOn(playerMock, 'getStreamType').mockReturnValue(StreamType.Dash);
       convivaAnalytics.updateContentMetadata({ custom: { playerType: PlayerType.Html5 }, assetName: 'MyAsset' });
@@ -77,13 +77,13 @@ describe('content metadata spec', () => {
 
       expect(convivaVideoAnalytics.reportPlaybackRequested).toHaveBeenLastCalledWith(
         expect.objectContaining({
-          playerType: PlayerType.Html5,
+          playerType: PlayerType.Native,
           streamType: StreamType.Dash,
         }),
       );
     });
 
-    it('override streamType in custom tags', () => {
+    it('do not override streamType in custom tags', () => {
       jest.spyOn(playerMock, 'getPlayerType').mockReturnValue(PlayerType.Native);
       jest.spyOn(playerMock, 'getStreamType').mockReturnValue(StreamType.Dash);
       convivaAnalytics.updateContentMetadata({ custom: { streamType: 'dash_vod' }, assetName: 'MyAsset' });
@@ -92,18 +92,18 @@ describe('content metadata spec', () => {
       expect(convivaVideoAnalytics.reportPlaybackRequested).toHaveBeenLastCalledWith(
         expect.objectContaining({
           playerType: PlayerType.Native,
-          streamType: 'dash_vod',
+          streamType: StreamType.Dash,
         }),
       );
     });
 
-    it('override vrContentType in custom tags', () => {
+    it('do not override vrContentType in custom tags', () => {
       convivaAnalytics.updateContentMetadata({ custom: { vrContentType: VRContentType.Single }, assetName: 'MyAsset' });
       playerMock.eventEmitter.firePlayEvent();
 
       expect(convivaVideoAnalytics.reportPlaybackRequested).toHaveBeenLastCalledWith(
         expect.objectContaining({
-          vrContentType: VRContentType.Single,
+          vrContentType: undefined,
         }),
       );
     });
@@ -178,6 +178,14 @@ describe('content metadata spec', () => {
             convivaAnalytics.initializeSession();
             expect(convivaVideoAnalytics.reportPlaybackRequested).toHaveBeenLastCalledWith(
               expect.objectContaining({ myTag: 'withMyValue' }),
+            );
+          });
+
+          it('additional standard tags', () => {
+            convivaAnalytics.updateContentMetadata({ additionalStandardTags: { 'c3.cm.brand': 'Test Brand' }, assetName: 'MyAsset' });
+            convivaAnalytics.initializeSession();
+            expect(convivaVideoAnalytics.reportPlaybackRequested).toHaveBeenLastCalledWith(
+              expect.objectContaining({ 'c3.cm.brand': 'Test Brand' }),
             );
           });
 
@@ -368,17 +376,23 @@ describe('content metadata spec', () => {
           it('custom tags', () => {
             convivaAnalytics.updateContentMetadata({ custom: { myTag: 'withMyValue' } });
 
-            expect(convivaVideoAnalytics.setContentInfo).not.toHaveBeenLastCalledWith(
-              0,
-              expect.objectContaining({ myTag: undefined }),
+            expect(convivaVideoAnalytics.setContentInfo).not.toHaveBeenCalledWith(
+              expect.objectContaining({ myTag: 'withMyValue' })
+            );
+          });
+
+          it('additional tags', () => {
+            convivaAnalytics.updateContentMetadata({ custom: { 'c3.cm.brand': 'Test Brand' } });
+
+            expect(convivaVideoAnalytics.setContentInfo).not.toHaveBeenCalledWith(
+              expect.objectContaining({ 'c3.cm.brand': 'Test Brand' })
             );
           });
 
           it('duration', () => {
             convivaAnalytics.updateContentMetadata({ duration: 55 });
 
-            expect(convivaVideoAnalytics.setContentInfo).not.toHaveBeenLastCalledWith(
-              0,
+            expect(convivaVideoAnalytics.setContentInfo).not.toHaveBeenCalledWith(
               expect.objectContaining({
                 duration: 55,
               }),